Enum input_linux::Key

#[repr(u16)]

pub enum Key {
    KeyReserved,
    KeyEsc,
    Key1,
    Key2,
    Key3,
    Key4,
    Key5,
    Key6,
    Key7,
    Key8,
    Key9,
    Key0,
    KeyMinus,
    KeyEqual,
    KeyBackspace,
    KeyTab,
    KeyQ,
    KeyW,
    KeyE,
    KeyR,
    KeyT,
    KeyY,
    KeyU,
    KeyI,
    KeyO,
    KeyP,
    KeyLeftBrace,
    KeyRightBrace,
    KeyEnter,
    KeyLeftCtrl,
    KeyA,
    KeyS,
    KeyD,
    KeyF,
    KeyG,
    KeyH,
    KeyJ,
    KeyK,
    KeyL,
    KeySemicolon,
    KeyApostrophe,
    KeyGrave,
    KeyLeftShift,
    KeyBackslash,
    KeyZ,
    KeyX,
    KeyC,
    KeyV,
    KeyB,
    KeyN,
    KeyM,
    KeyComma,
    KeyDot,
    KeySlash,
    KeyRightShift,
    KeyKpAsterisk,
    KeyLeftAlt,
    KeySpace,
    KeyCapsLock,
    KeyF1,
    KeyF2,
    KeyF3,
    KeyF4,
    KeyF5,
    KeyF6,
    KeyF7,
    KeyF8,
    KeyF9,
    KeyF10,
    KeyNumLock,
    KeyScrollLock,
    KeyKp7,
    KeyKp8,
    KeyKp9,
    KeyKpMinus,
    KeyKp4,
    KeyKp5,
    KeyKp6,
    KeyKpPlus,
    KeyKp1,
    KeyKp2,
    KeyKp3,
    KeyKp0,
    KeyKpDot,
    Unknown54,
    KeyZenkakuHankaku,
    Key102nd,
    KeyF11,
    KeyF12,
    KeyRo,
    KeyKatakana,
    KeyHiragana,
    KeyHenkan,
    KeyKatakanaHiragana,
    KeyMuhenkan,
    KeyKpJpComma,
    KeyKpEnter,
    KeyRightCtrl,
    KeyKpSlash,
    KeySysrq,
    KeyRightAlt,
    KeyLineFeed,
    KeyHome,
    KeyUp,
    KeyPageUp,
    KeyLeft,
    KeyRight,
    KeyEnd,
    KeyDown,
    KeyPageDown,
    KeyInsert,
    KeyDelete,
    KeyMacro,
    KeyMute,
    KeyVolumeDown,
    KeyVolumeUp,
    KeyPower,
    KeyKpEqual,
    KeyKpPlusMinus,
    KeyPause,
    KeyScale,
    KeyKpComma,
    KeyHangul,
    KeyHanja,
    KeyYen,
    KeyLeftMeta,
    KeyRightMeta,
    KeyCompose,
    KeyStop,
    KeyAgain,
    KeyProps,
    KeyUndo,
    KeyFront,
    KeyCopy,
    KeyOpen,
    KeyPaste,
    KeyFind,
    KeyCut,
    KeyHelp,
    KeyMenu,
    KeyCalc,
    KeySetup,
    KeySleep,
    KeyWakeup,
    KeyFile,
    KeySendFile,
    KeyDeleteFile,
    KeyXfer,
    KeyProg1,
    KeyProg2,
    KeyWWW,
    KeyMSDOS,
    KeyCoffee,
    KeyRotateDisplay,
    KeyCycleWindows,
    KeyMail,
    KeyBookmarks,
    KeyComputer,
    KeyBack,
    KeyForward,
    KeyCloseCD,
    KeyEjectCD,
    KeyEjectCloseCD,
    KeyNextSong,
    KeyPlayPause,
    KeyPreviousSong,
    KeyStopCD,
    KeyRecord,
    KeyRewind,
    KeyPhone,
    KeyIso,
    KeyConfig,
    KeyHomepage,
    KeyRefresh,
    KeyExit,
    KeyMove,
    KeyEdit,
    KeyScrollUp,
    KeyScrollDown,
    KeyKpLeftParen,
    KeyKpRightParen,
    KeyNew,
    KeyRedo,
    KeyF13,
    KeyF14,
    KeyF15,
    KeyF16,
    KeyF17,
    KeyF18,
    KeyF19,
    KeyF20,
    KeyF21,
    KeyF22,
    KeyF23,
    KeyF24,
    UnknownC3,
    UnknownC4,
    UnknownC5,
    UnknownC6,
    UnknownC7,
    KeyPlayCD,
    KeyPauseCD,
    KeyProg3,
    KeyProg4,
    KeyDashboard,
    KeySuspend,
    KeyClose,
    KeyPlay,
    KeyFastForward,
    KeyBassBoost,
    KeyPrint,
    KeyHp,
    KeyCamera,
    KeySound,
    KeyQuestion,
    KeyEmail,
    KeyChat,
    KeySearch,
    KeyConnect,
    KeyFinance,
    KeySport,
    KeyShop,
    KeyAlterase,
    KeyCancel,
    KeyBrightnessDown,
    KeyBrightnessUp,
    KeyMedia,
    KeySwitchVideoMode,
    KeyIllumToggle,
    KeyIllumDown,
    KeyIllumUp,
    KeySend,
    KeyReply,
    KeyForwardMail,
    KeySave,
    KeyDocuments,
    KeyBattery,
    KeyBluetooth,
    KeyWLAN,
    KeyUWB,
    KeyUnknown,
    KeyVideoNext,
    KeyVideoPrev,
    KeyBrightnessCycle,
    KeyBrightnessAuto,
    KeyDisplayOff,
    KeyWWAN,
    KeyRfkill,
    KeyMicMute,
    UnknownF9,
    UnknownFA,
    UnknownFB,
    UnknownFC,
    UnknownFD,
    UnknownFE,
    ReservedFF,
    Button0,
    Button1,
    Button2,
    Button3,
    Button4,
    Button5,
    Button6,
    Button7,
    Button8,
    Button9,
    Unknown10A,
    Unknown10B,
    Unknown10C,
    Unknown10D,
    Unknown10E,
    Unknown10F,
    ButtonLeft,
    ButtonRight,
    ButtonMiddle,
    ButtonSide,
    ButtonExtra,
    ButtonForward,
    ButtonBack,
    ButtonTask,
    Unknown118,
    Unknown119,
    Unknown11A,
    Unknown11B,
    Unknown11C,
    Unknown11D,
    Unknown11E,
    Unknown11F,
    ButtonTrigger,
    ButtonThumb,
    ButtonThumb2,
    ButtonTop,
    ButtonTop2,
    ButtonPinkie,
    ButtonBase,
    ButtonBase2,
    ButtonBase3,
    ButtonBase4,
    ButtonBase5,
    ButtonBase6,
    Unknown12C,
    Unknown12D,
    Unknown12E,
    ButtonDead,
    ButtonSouth,
    ButtonEast,
    ButtonC,
    ButtonNorth,
    ButtonWest,
    ButtonZ,
    ButtonTL,
    ButtonTR,
    ButtonTL2,
    ButtonTR2,
    ButtonSelect,
    ButtonStart,
    ButtonMode,
    ButtonThumbl,
    ButtonThumbr,
    Unknown13F,
    ButtonToolPen,
    ButtonToolRubber,
    ButtonToolBrush,
    ButtonToolPencil,
    ButtonToolAirbrush,
    ButtonToolFinger,
    ButtonToolMouse,
    ButtonToolLens,
    ButtonToolQuintTap,
    Unknown149,
    ButtonTouch,
    ButtonStylus,
    ButtonStylus2,
    ButtonToolDoubleTap,
    ButtonToolTripleTap,
    ButtonToolQuadtap,
    ButtonWheel,
    ButtonGearUp,
    Unknown152,
    Unknown153,
    Unknown154,
    Unknown155,
    Unknown156,
    Unknown157,
    Unknown158,
    Unknown159,
    Unknown15A,
    Unknown15B,
    Unknown15C,
    Unknown15D,
    Unknown15E,
    Unknown15F,
    KeyOk,
    KeySelect,
    KeyGoto,
    KeyClear,
    KeyPower2,
    KeyOption,
    KeyInfo,
    KeyTime,
    KeyVendor,
    KeyArchive,
    KeyProgram,
    KeyChannel,
    KeyFavorites,
    KeyEPG,
    KeyPVR,
    KeyMHP,
    KeyLanguage,
    KeyTitle,
    KeySubtitle,
    KeyAngle,
    KeyZoom,
    KeyMode,
    KeyKeyboard,
    KeyScreen,
    KeyPC,
    KeyTV,
    KeyTV2,
    KeyVCR,
    KeyVCR2,
    KeySat,
    KeySat2,
    KeyCD,
    KeyTape,
    KeyRadio,
    KeyTuner,
    KeyPlayer,
    KeyText,
    KeyDvd,
    KeyAux,
    KeyMp3,
    KeyAudio,
    KeyVideo,
    KeyDirectory,
    KeyList,
    KeyMemo,
    KeyCalendar,
    KeyRed,
    KeyGreen,
    KeyYellow,
    KeyBlue,
    KeyChannelUp,
    KeyChannelDown,
    KeyFirst,
    KeyLast,
    KeyAb,
    KeyNext,
    KeyRestart,
    KeySlow,
    KeyShuffle,
    KeyBreak,
    KeyPrevious,
    KeyDigits,
    KeyTeen,
    KeyTwen,
    KeyVideophone,
    KeyGames,
    KeyZoomIn,
    KeyZoomOut,
    KeyZoomReset,
    KeyWordProcessor,
    KeyEditor,
    KeySpreadsheet,
    KeyGraphicsEditor,
    KeyPresentation,
    KeyDatabase,
    KeyNews,
    KeyVoicemail,
    KeyAddressBook,
    KeyMessenger,
    KeyDisplayToggle,
    KeySpellCheck,
    KeyLogoff,
    KeyDollar,
    KeyEuro,
    KeyFrameBack,
    KeyFrameForward,
    KeyContextMenu,
    KeyMediaRepeat,
    Key10ChannelsUp,
    Key10ChannelsDown,
    KeyImages,
    Unknown1BB,
    Unknown1BC,
    Unknown1BD,
    Unknown1BE,
    Unknown1BF,
    KeyDelEol,
    KeyDelEos,
    KeyInsLine,
    KeyDelLine,
    Unknown1C4,
    Unknown1C5,
    Unknown1C6,
    Unknown1C7,
    Unknown1C8,
    Unknown1C9,
    Unknown1CA,
    Unknown1CB,
    Unknown1CC,
    Unknown1CD,
    Unknown1CE,
    Unknown1CF,
    KeyFn,
    KeyFnEsc,
    KeyFnF1,
    KeyFnF2,
    KeyFnF3,
    KeyFnF4,
    KeyFnF5,
    KeyFnF6,
    KeyFnF7,
    KeyFnF8,
    KeyFnF9,
    KeyFnF10,
    KeyFnF11,
    KeyFnF12,
    KeyFn1,
    KeyFn2,
    KeyFnD,
    KeyFnE,
    KeyFnF,
    KeyFnS,
    KeyFnB,
    Unknown1E5,
    Unknown1E6,
    Unknown1E7,
    Unknown1E8,
    Unknown1E9,
    Unknown1EA,
    Unknown1EB,
    Unknown1EC,
    Unknown1ED,
    Unknown1EE,
    Unknown1EF,
    Unknown1F0,
    KeyBrlDot1,
    KeyBrlDot2,
    KeyBrlDot3,
    KeyBrlDot4,
    KeyBrlDot5,
    KeyBrlDot6,
    KeyBrlDot7,
    KeyBrlDot8,
    KeyBrlDot9,
    KeyBrlDot10,
    Unknown1FB,
    Unknown1FC,
    Unknown1FD,
    Unknown1FE,
    Unknown1FF,
    KeyNumeric0,
    KeyNumeric1,
    KeyNumeric2,
    KeyNumeric3,
    KeyNumeric4,
    KeyNumeric5,
    KeyNumeric6,
    KeyNumeric7,
    KeyNumeric8,
    KeyNumeric9,
    KeyNumericStar,
    KeyNumericPound,
    KeyNumericA,
    KeyNumericB,
    KeyNumericC,
    KeyNumericD,
    KeyCameraFocus,
    KeyWpsButton,
    KeyTouchpadToggle,
    KeyTouchpadOn,
    KeyTouchpadOff,
    KeyCameraZoomin,
    KeyCameraZoomout,
    KeyCameraUp,
    KeyCameraDown,
    KeyCameraLeft,
    KeyCameraRight,
    KeyAttendantOn,
    KeyAttendantOff,
    KeyAttendantToggle,
    KeyLightsToggle,
    Unknown21F,
    ButtonDpadUp,
    ButtonDpadDown,
    ButtonDpadLeft,
    ButtonDpadRight,
    Unknown224,
    Unknown225,
    Unknown226,
    Unknown227,
    Unknown228,
    Unknown229,
    Unknown22A,
    Unknown22B,
    Unknown22C,
    Unknown22D,
    Unknown22E,
    Unknown22F,
    KeyAlsToggle,
    Unknown231,
    Unknown232,
    Unknown233,
    Unknown234,
    Unknown235,
    Unknown236,
    Unknown237,
    Unknown238,
    Unknown239,
    Unknown23A,
    Unknown23B,
    Unknown23C,
    Unknown23D,
    Unknown23E,
    Unknown23F,
    KeyButtonConfig,
    KeyTaskManager,
    KeyJournal,
    KeyControlPanel,
    KeyAppSelect,
    KeyScreensaver,
    KeyVoicecommand,
    Unknown247,
    Unknown248,
    Unknown249,
    Unknown24A,
    Unknown24B,
    Unknown24C,
    Unknown24D,
    Unknown24E,
    Unknown24F,
    KeyBrightnessMin,
    KeyBrightnessMax,
    Unknown252,
    Unknown253,
    Unknown254,
    Unknown255,
    Unknown256,
    Unknown257,
    Unknown258,
    Unknown259,
    Unknown25A,
    Unknown25B,
    Unknown25C,
    Unknown25D,
    Unknown25E,
    Unknown25F,
    KeyInputAssistPrev,
    KeyInputAssistNext,
    KeyInputAssistPrevGroup,
    KeyInputAssistNextGroup,
    KeyInputAssistAccept,
    KeyInputAssistCancel,
    KeyRightUp,
    KeyRightDown,
    KeyLeftUp,
    KeyLeftDown,
    KeyRootMenu,
    KeyMediaTopMenu,
    KeyNumeric11,
    KeyNumeric12,
    KeyAudioDesc,
    Key3dMode,
    KeyNextFavorite,
    KeyStopRecord,
    KeyPauseRecord,
    KeyVod,
    KeyUnmute,
    KeyFastReverse,
    KeySlowReverse,
    KeyData,
    KeyOnscreenKeyboard,
    Unknown279,
    Unknown27A,
    Unknown27B,
    Unknown27C,
    Unknown27D,
    Unknown27E,
    Unknown27F,
    Unknown280,
    Unknown281,
    Unknown282,
    Unknown283,
    Unknown284,
    Unknown285,
    Unknown286,
    Unknown287,
    Unknown288,
    Unknown289,
    Unknown28A,
    Unknown28B,
    Unknown28C,
    Unknown28D,
    Unknown28E,
    Unknown28F,
    Unknown290,
    Unknown291,
    Unknown292,
    Unknown293,
    Unknown294,
    Unknown295,
    Unknown296,
    Unknown297,
    Unknown298,
    Unknown299,
    Unknown29A,
    Unknown29B,
    Unknown29C,
    Unknown29D,
    Unknown29E,
    Unknown29F,
    Unknown2A0,
    Unknown2A1,
    Unknown2A2,
    Unknown2A3,
    Unknown2A4,
    Unknown2A5,
    Unknown2A6,
    Unknown2A7,
    Unknown2A8,
    Unknown2A9,
    Unknown2AA,
    Unknown2AB,
    Unknown2AC,
    Unknown2AD,
    Unknown2AE,
    Unknown2AF,
    Unknown2B0,
    Unknown2B1,
    Unknown2B2,
    Unknown2B3,
    Unknown2B4,
    Unknown2B5,
    Unknown2B6,
    Unknown2B7,
    Unknown2B8,
    Unknown2B9,
    Unknown2BA,
    Unknown2BB,
    Unknown2BC,
    Unknown2BD,
    Unknown2BE,
    Unknown2BF,
    ButtonTriggerHappy1,
    ButtonTriggerHappy2,
    ButtonTriggerHappy3,
    ButtonTriggerHappy4,
    ButtonTriggerHappy5,
    ButtonTriggerHappy6,
    ButtonTriggerHappy7,
    ButtonTriggerHappy8,
    ButtonTriggerHappy9,
    ButtonTriggerHappy10,
    ButtonTriggerHappy11,
    ButtonTriggerHappy12,
    ButtonTriggerHappy13,
    ButtonTriggerHappy14,
    ButtonTriggerHappy15,
    ButtonTriggerHappy16,
    ButtonTriggerHappy17,
    ButtonTriggerHappy18,
    ButtonTriggerHappy19,
    ButtonTriggerHappy20,
    ButtonTriggerHappy21,
    ButtonTriggerHappy22,
    ButtonTriggerHappy23,
    ButtonTriggerHappy24,
    ButtonTriggerHappy25,
    ButtonTriggerHappy26,
    ButtonTriggerHappy27,
    ButtonTriggerHappy28,
    ButtonTriggerHappy29,
    ButtonTriggerHappy30,
    ButtonTriggerHappy31,
    ButtonTriggerHappy32,
    ButtonTriggerHappy33,
    ButtonTriggerHappy34,
    ButtonTriggerHappy35,
    ButtonTriggerHappy36,
    ButtonTriggerHappy37,
    ButtonTriggerHappy38,
    ButtonTriggerHappy39,
    ButtonTriggerHappy40,
    Unknown2E8,
    Unknown2E9,
    Unknown2EA,
    Unknown2EB,
    Unknown2EC,
    Unknown2ED,
    Unknown2EE,
    Unknown2EF,
    Unknown2F0,
    Unknown2F1,
    Unknown2F2,
    Unknown2F3,
    Unknown2F4,
    Unknown2F5,
    Unknown2F6,
    Unknown2F7,
    Unknown2F8,
    Unknown2F9,
    Unknown2FA,
    Unknown2FB,
    Unknown2FC,
    Unknown2FD,
    Unknown2FE,
    Unknown2FF,
}

Keys and Buttons

Most of the keys/buttons are modeled after USB HUT 1.12 (see http://www.usb.org/developers/hidpage).

Comment Abbreviations

AC - Application Control AL - Application Launch Button SC - System Control

Variants
KeyReservedKeyEscKey1Key2Key3Key4Key5Key6Key7Key8Key9Key0KeyMinusKeyEqualKeyBackspaceKeyTabKeyQKeyWKeyEKeyRKeyTKeyYKeyUKeyIKeyOKeyPKeyLeftBraceKeyRightBraceKeyEnterKeyLeftCtrlKeyAKeySKeyDKeyFKeyGKeyHKeyJKeyKKeyLKeySemicolonKeyApostropheKeyGraveKeyLeftShiftKeyBackslashKeyZKeyXKeyCKeyVKeyBKeyNKeyMKeyCommaKeyDotKeySlashKeyRightShiftKeyKpAsteriskKeyLeftAltKeySpaceKeyCapsLockKeyF1KeyF2KeyF3KeyF4KeyF5KeyF6KeyF7KeyF8KeyF9KeyF10KeyNumLockKeyScrollLockKeyKp7KeyKp8KeyKp9KeyKpMinusKeyKp4KeyKp5KeyKp6KeyKpPlusKeyKp1KeyKp2KeyKp3KeyKp0KeyKpDotUnknown54KeyZenkakuHankakuKey102ndKeyF11KeyF12KeyRoKeyKatakanaKeyHiraganaKeyHenkanKeyKatakanaHiraganaKeyMuhenkanKeyKpJpCommaKeyKpEnterKeyRightCtrlKeyKpSlashKeySysrqKeyRightAltKeyLineFeedKeyHomeKeyUpKeyPageUpKeyLeftKeyRightKeyEndKeyDownKeyPageDownKeyInsertKeyDeleteKeyMacroKeyMuteKeyVolumeDownKeyVolumeUpKeyPower

SC System Power Down

KeyKpEqualKeyKpPlusMinusKeyPauseKeyScale

AL Compiz Scale (Expose)

KeyKpCommaKeyHangul

KeyHangeul / KeyHanguel

KeyHanjaKeyYenKeyLeftMetaKeyRightMetaKeyComposeKeyStop

AC Stop

KeyAgainKeyProps

AC Properties

KeyUndo

AC Undo

KeyFrontKeyCopy

AC Copy

KeyOpen

AC Open

KeyPaste

AC Paste

KeyFind

AC Search

KeyCut

AC Cut

KeyHelp

AL Integrated Help Center

KeyMenu

Menu (show menu)

KeyCalc

AL Calculator

KeySetupKeySleep

SC System Sleep

KeyWakeup

System Wake Up

KeyFile

AL Local Machine Browser

KeySendFileKeyDeleteFileKeyXferKeyProg1KeyProg2KeyWWW

AL Internet Browser

KeyMSDOSKeyCoffee

AL Terminal Lock/Screensaver KeyScreenLock

KeyRotateDisplay

Display orientation for e.g. tablets (aka KeyDirectionKey)

KeyCycleWindowsKeyMailKeyBookmarks

AC Bookmarks

KeyComputerKeyBack

AC Back

KeyForward

AC Forward

KeyCloseCDKeyEjectCDKeyEjectCloseCDKeyNextSongKeyPlayPauseKeyPreviousSongKeyStopCDKeyRecordKeyRewindKeyPhone

Media Select Telephone

KeyIsoKeyConfig

AL Consumer Control Configuration

KeyHomepage

AC Home

KeyRefresh

AC Refresh

KeyExit

AC Exit

KeyMoveKeyEditKeyScrollUpKeyScrollDownKeyKpLeftParenKeyKpRightParenKeyNew

AC New

KeyRedo

AC Redo/Repeat

KeyF13KeyF14KeyF15KeyF16KeyF17KeyF18KeyF19KeyF20KeyF21KeyF22KeyF23KeyF24UnknownC3UnknownC4UnknownC5UnknownC6UnknownC7KeyPlayCDKeyPauseCDKeyProg3KeyProg4KeyDashboard

AL Dashboard

KeySuspendKeyClose

AC Close

KeyPlayKeyFastForwardKeyBassBoostKeyPrint

AC Print

KeyHpKeyCameraKeySoundKeyQuestionKeyEmailKeyChatKeySearchKeyConnectKeyFinance

AL Checkbook/Finance

KeySportKeyShopKeyAlteraseKeyCancel

AC Cancel

KeyBrightnessDownKeyBrightnessUpKeyMediaKeySwitchVideoMode

Cycle between available video outputs (Monitor/LCD/TV-out/etc)

KeyIllumToggleKeyIllumDownKeyIllumUpKeySend

AC Send

KeyReply

AC Reply

KeyForwardMail

AC Forward Msg

KeySave

AC Save

KeyDocumentsKeyBatteryKeyBluetoothKeyWLANKeyUWBKeyUnknownKeyVideoNext

drive next video source

KeyVideoPrev

drive previous video source

KeyBrightnessCycle

brightness up, after max is min

KeyBrightnessAuto

Set Auto Brightness: manual brightness control is off, rely on ambient (aka KeyBrightnessZero)

KeyDisplayOff

display device to off state

KeyWWAN

Wireless WAN (LTE, UMTS, GSM, etc.) (aka KeyWiMAX)

KeyRfkill

Key that controls all radios

KeyMicMute

Mute / unmute the microphone

UnknownF9UnknownFAUnknownFBUnknownFCUnknownFDUnknownFEReservedFF

Code 255 is reserved for special needs of AT keyboard driver

Button0Button1Button2Button3Button4Button5Button6Button7Button8Button9Unknown10AUnknown10BUnknown10CUnknown10DUnknown10EUnknown10FButtonLeftButtonRightButtonMiddleButtonSideButtonExtraButtonForwardButtonBackButtonTaskUnknown118Unknown119Unknown11AUnknown11BUnknown11CUnknown11DUnknown11EUnknown11FButtonTriggerButtonThumbButtonThumb2ButtonTopButtonTop2ButtonPinkieButtonBaseButtonBase2ButtonBase3ButtonBase4ButtonBase5ButtonBase6Unknown12CUnknown12DUnknown12EButtonDeadButtonSouth

aka ButtonA

ButtonEast

aka ButtonB

ButtonCButtonNorth

aka ButtonX

ButtonWest

aka ButtonY

ButtonZButtonTLButtonTRButtonTL2ButtonTR2ButtonSelectButtonStartButtonModeButtonThumblButtonThumbrUnknown13FButtonToolPenButtonToolRubberButtonToolBrushButtonToolPencilButtonToolAirbrushButtonToolFingerButtonToolMouseButtonToolLensButtonToolQuintTap

Five fingers on trackpad

Unknown149ButtonTouchButtonStylusButtonStylus2ButtonToolDoubleTapButtonToolTripleTapButtonToolQuadtap

Four fingers on trackpad

ButtonWheelButtonGearUpUnknown152Unknown153Unknown154Unknown155Unknown156Unknown157Unknown158Unknown159Unknown15AUnknown15BUnknown15CUnknown15DUnknown15EUnknown15FKeyOkKeySelectKeyGotoKeyClearKeyPower2KeyOptionKeyInfo

AL OEM Features/Tips/Tutorial

KeyTimeKeyVendorKeyArchiveKeyProgram

Media Select Program Guide

KeyChannelKeyFavoritesKeyEPGKeyPVR

Media Select Home

KeyMHPKeyLanguageKeyTitleKeySubtitleKeyAngleKeyZoomKeyModeKeyKeyboardKeyScreenKeyPC

Media Select Computer

KeyTV

Media Select TV

KeyTV2

Media Select Cable

KeyVCR

Media Select VCR

KeyVCR2

VCR Plus

KeySat

Media Select Satellite

KeySat2KeyCD

Media Select CD

KeyTape

Media Select Tape

KeyRadioKeyTuner

Media Select Tuner

KeyPlayerKeyTextKeyDvd

Media Select DVD

KeyAuxKeyMp3KeyAudio

AL Audio Browser

KeyVideo

AL Movie Browser

KeyDirectoryKeyListKeyMemo

Media Select Messages

KeyCalendarKeyRedKeyGreenKeyYellowKeyBlueKeyChannelUp

Channel Increment

KeyChannelDown

Channel Decrement

KeyFirstKeyLast

Recall Last

KeyAbKeyNextKeyRestartKeySlowKeyShuffleKeyBreakKeyPreviousKeyDigitsKeyTeenKeyTwenKeyVideophone

Media Select Video Phone

KeyGames

Media Select Games

KeyZoomIn

AC Zoom In

KeyZoomOut

AC Zoom Out

KeyZoomReset

AC Zoom

KeyWordProcessor

AL Word Processor

KeyEditor

AL Text Editor

KeySpreadsheet

AL Spreadsheet

KeyGraphicsEditor

AL Graphics Editor

KeyPresentation

AL Presentation App

KeyDatabase

AL Database App

KeyNews

AL Newsreader

KeyVoicemail

AL Voicemail

KeyAddressBook

AL Contacts/Address Book

KeyMessenger

AL Instant Messaging

KeyDisplayToggle

Turn display (LCD) on and off (aka KeyBrightnessToggle)

KeySpellCheck

AL Spell Check

KeyLogoff

AL Logoff

KeyDollarKeyEuroKeyFrameBack

Consumer - transport controls

KeyFrameForwardKeyContextMenu

GenDesc - system context menu

KeyMediaRepeat

Consumer - transport control

Key10ChannelsUp

10 channels up (10+)

Key10ChannelsDown

10 channels down (10-)

KeyImages

AL Image Browser

Unknown1BBUnknown1BCUnknown1BDUnknown1BEUnknown1BFKeyDelEolKeyDelEosKeyInsLineKeyDelLineUnknown1C4Unknown1C5Unknown1C6Unknown1C7Unknown1C8Unknown1C9Unknown1CAUnknown1CBUnknown1CCUnknown1CDUnknown1CEUnknown1CFKeyFnKeyFnEscKeyFnF1KeyFnF2KeyFnF3KeyFnF4KeyFnF5KeyFnF6KeyFnF7KeyFnF8KeyFnF9KeyFnF10KeyFnF11KeyFnF12KeyFn1KeyFn2KeyFnDKeyFnEKeyFnFKeyFnSKeyFnBUnknown1E5Unknown1E6Unknown1E7Unknown1E8Unknown1E9Unknown1EAUnknown1EBUnknown1ECUnknown1EDUnknown1EEUnknown1EFUnknown1F0KeyBrlDot1KeyBrlDot2KeyBrlDot3KeyBrlDot4KeyBrlDot5KeyBrlDot6KeyBrlDot7KeyBrlDot8KeyBrlDot9KeyBrlDot10Unknown1FBUnknown1FCUnknown1FDUnknown1FEUnknown1FFKeyNumeric0

used by phones, remote controls,

KeyNumeric1

and other keypads

KeyNumeric2KeyNumeric3KeyNumeric4KeyNumeric5KeyNumeric6KeyNumeric7KeyNumeric8KeyNumeric9KeyNumericStarKeyNumericPoundKeyNumericA

Phone key A - HUT Telephony 0xb9

KeyNumericBKeyNumericCKeyNumericDKeyCameraFocusKeyWpsButton

WiFi Protected Setup key

KeyTouchpadToggle

Request switch touchpad on or off

KeyTouchpadOnKeyTouchpadOffKeyCameraZoominKeyCameraZoomoutKeyCameraUpKeyCameraDownKeyCameraLeftKeyCameraRightKeyAttendantOnKeyAttendantOffKeyAttendantToggle

Attendant call on or off

KeyLightsToggle

Reading light on or off

Unknown21FButtonDpadUpButtonDpadDownButtonDpadLeftButtonDpadRightUnknown224Unknown225Unknown226Unknown227Unknown228Unknown229Unknown22AUnknown22BUnknown22CUnknown22DUnknown22EUnknown22FKeyAlsToggle

Ambient light sensor

Unknown231Unknown232Unknown233Unknown234Unknown235Unknown236Unknown237Unknown238Unknown239Unknown23AUnknown23BUnknown23CUnknown23DUnknown23EUnknown23FKeyButtonConfig

AL Button Configuration

KeyTaskManager

AL Task/Project Manager

KeyJournal

AL Log/Journal/Timecard

KeyControlPanel

AL Control Panel

KeyAppSelect

AL Select Task/Application

KeyScreensaver

AL Screen Saver

KeyVoicecommand

Listening Voice Command

Unknown247Unknown248Unknown249Unknown24AUnknown24BUnknown24CUnknown24DUnknown24EUnknown24FKeyBrightnessMin

Set Brightness to Minimum

KeyBrightnessMax

Set Brightness to Maximum

Unknown252Unknown253Unknown254Unknown255Unknown256Unknown257Unknown258Unknown259Unknown25AUnknown25BUnknown25CUnknown25DUnknown25EUnknown25FKeyInputAssistPrevKeyInputAssistNextKeyInputAssistPrevGroupKeyInputAssistNextGroupKeyInputAssistAcceptKeyInputAssistCancelKeyRightUp

Diagonal movement keys

KeyRightDownKeyLeftUpKeyLeftDownKeyRootMenu

Show Device's Root Menu

KeyMediaTopMenu

Show Top Menu of the Media (e.g. DVD)

KeyNumeric11KeyNumeric12KeyAudioDesc

Toggle Audio Description: refers to an audio service that helps blind and visually impaired consumers understand the action in a program. Note: in some countries this is referred to as "Video Description".

Key3dModeKeyNextFavoriteKeyStopRecordKeyPauseRecordKeyVod

Video on Demand

KeyUnmuteKeyFastReverseKeySlowReverseKeyData

Control a data application associated with the currently viewed channel, e.g. teletext or data broadcast application (MHEG, MHP, HbbTV, etc.)

KeyOnscreenKeyboardUnknown279Unknown27AUnknown27BUnknown27CUnknown27DUnknown27EUnknown27FUnknown280Unknown281Unknown282Unknown283Unknown284Unknown285Unknown286Unknown287Unknown288Unknown289Unknown28AUnknown28BUnknown28CUnknown28DUnknown28EUnknown28FUnknown290Unknown291Unknown292Unknown293Unknown294Unknown295Unknown296Unknown297Unknown298Unknown299Unknown29AUnknown29BUnknown29CUnknown29DUnknown29EUnknown29FUnknown2A0Unknown2A1Unknown2A2Unknown2A3Unknown2A4Unknown2A5Unknown2A6Unknown2A7Unknown2A8Unknown2A9Unknown2AAUnknown2ABUnknown2ACUnknown2ADUnknown2AEUnknown2AFUnknown2B0Unknown2B1Unknown2B2Unknown2B3Unknown2B4Unknown2B5Unknown2B6Unknown2B7Unknown2B8Unknown2B9Unknown2BAUnknown2BBUnknown2BCUnknown2BDUnknown2BEUnknown2BFButtonTriggerHappy1ButtonTriggerHappy2ButtonTriggerHappy3ButtonTriggerHappy4ButtonTriggerHappy5ButtonTriggerHappy6ButtonTriggerHappy7ButtonTriggerHappy8ButtonTriggerHappy9ButtonTriggerHappy10ButtonTriggerHappy11ButtonTriggerHappy12ButtonTriggerHappy13ButtonTriggerHappy14ButtonTriggerHappy15ButtonTriggerHappy16ButtonTriggerHappy17ButtonTriggerHappy18ButtonTriggerHappy19ButtonTriggerHappy20ButtonTriggerHappy21ButtonTriggerHappy22ButtonTriggerHappy23ButtonTriggerHappy24ButtonTriggerHappy25ButtonTriggerHappy26ButtonTriggerHappy27ButtonTriggerHappy28ButtonTriggerHappy29ButtonTriggerHappy30ButtonTriggerHappy31ButtonTriggerHappy32ButtonTriggerHappy33ButtonTriggerHappy34ButtonTriggerHappy35ButtonTriggerHappy36ButtonTriggerHappy37ButtonTriggerHappy38ButtonTriggerHappy39ButtonTriggerHappy40Unknown2E8Unknown2E9Unknown2EAUnknown2EBUnknown2ECUnknown2EDUnknown2EEUnknown2EFUnknown2F0Unknown2F1Unknown2F2Unknown2F3Unknown2F4Unknown2F5Unknown2F6Unknown2F7Unknown2F8Unknown2F9Unknown2FAUnknown2FBUnknown2FCUnknown2FDUnknown2FEUnknown2FF
Methods

impl Key

pub const COUNT: usize

COUNT: usize = sys::KEY_CNT as usize

Important traits for EnumIterator<T>
impl<T: IterableEnum> Iterator for EnumIterator<T> type Item = T;

pub fn iter() -> EnumIterator<Self>

pub fn bitmask() -> Bitmask<Self>

impl Key

pub fn from_code(code: u16) -> Result<Self, RangeError>

impl Key

pub fn is_button(&self) -> bool

pub fn is_key(&self) -> bool