Struct input_linux::evdev::EvdevHandle

pub struct EvdevHandle(_);

A handle to an input device allowing the use of ioctls

Ownership of the file descriptor is not transferred, and it must stay open for this object's lifetime. It will not be closed automatically.

Methods

impl EvdevHandle

pub fn new<F: AsRawFd>(fd: &F) -> Self

Create a new handle using an existing open file object.

pub fn from_fd(fd: RawFd) -> Self

Create a new handle from a raw file descriptor.

pub fn read(&self, events: &mut [input_event]) -> Result<usize>

Read events from the input device

pub fn driver_version(&self) -> Result<i32>

EVIOCGVERSION

pub fn device_id(&self) -> Result<InputId>

EVIOCGID

pub fn repeat_settings(&self) -> Result<repeat_settings>

EVIOGREP

pub fn set_repeat_settings(&self, value: &repeat_settings) -> Result<()>

EVIOSREP

pub fn keycode_legacy(&self) -> Result<[u32; 2]>

EVIOCGKEYCODE

pub fn keycode(&self) -> Result<input_keymap_entry>

EVIOCGKEYCODE_V2

pub fn set_keycode_legacy(&self, value: &[u32; 2]) -> Result<()>

EVIOCSKEYCODE

pub fn set_keycode(&self, value: &input_keymap_entry) -> Result<()>

EVIOCSKEYCODE_V2

pub fn device_name_buf(&self, buffer: &mut [u8]) -> Result<usize>

EVIOCGNAME

pub fn device_name(&self) -> Result<Vec<u8>>

EVIOCGNAME

pub fn physical_location_buf(&self, buffer: &mut [u8]) -> Result<usize>

EVIOCGPHYS

pub fn physical_location(&self) -> Result<Vec<u8>>

EVIOCGPHYS

pub fn unique_id_buf(&self, buffer: &mut [u8]) -> Result<usize>

EVIOCGUNIQ

pub fn unique_id(&self) -> Result<Vec<u8>>

EVIOCGUNIQ

pub fn device_properties_raw(&self, buffer: &mut [u8]) -> Result<usize>

EVIOCGPROP

pub fn key_state(&self, buffer: &mut [u8]) -> Result<usize>

EVIOCGKEY

pub fn led_state(&self, buffer: &mut [u8]) -> Result<usize>

EVIOCGLED

pub fn sounds_state(&self, buffer: &mut [u8]) -> Result<usize>

EVIOCGSND

pub fn switch_state(&self, buffer: &mut [u8]) -> Result<usize>

EVIOCGSW

pub fn send_force_feedback(&self, value: &mut ff_effect) -> Result<()>

EVIOCSFF

pub fn erase_force_feedback(&self, value: i16) -> Result<()>

EVIOCRMFF

pub fn effects_count(&self) -> Result<i32>

EVIOCGEFFECTS

pub fn device_properties(&self) -> Result<Bitmask<InputProperty>>

EVIOCGPROP

pub fn multi_touch_slots(
&self,
code: AbsoluteAxis,
values: &mut [i32]
) -> Result<()>

EVIOCGMTSLOTS

Warning: the current implementation can leak uninitialized heap memory into values

pub fn event_mask_events(&self) -> Result<Bitmask<EventKind>>

EVIOCGMASK

pub fn set_event_mask_events(&self, bitmask: &Bitmask<EventKind>) -> Result<()>

EVIOCSMASK

pub fn event_bits(&self) -> Result<Bitmask<EventKind>>

EVIOCGBIT

pub fn key_mask(&self) -> Result<Bitmask<Key>>

EVIOCGMASK

pub fn set_key_mask(&self, bitmask: &Bitmask<Key>) -> Result<()>

EVIOCSMASK

pub fn key_bits(&self) -> Result<Bitmask<Key>>

EVIOCGBIT

pub fn relative_mask(&self) -> Result<Bitmask<RelativeAxis>>

EVIOCGMASK

pub fn set_relative_mask(&self, bitmask: &Bitmask<RelativeAxis>) -> Result<()>

EVIOCSMASK

pub fn relative_bits(&self) -> Result<Bitmask<RelativeAxis>>

EVIOCGBIT

pub fn absolute_mask(&self) -> Result<Bitmask<AbsoluteAxis>>

EVIOCGMASK

pub fn set_absolute_mask(&self, bitmask: &Bitmask<AbsoluteAxis>) -> Result<()>

EVIOCSMASK

pub fn absolute_bits(&self) -> Result<Bitmask<AbsoluteAxis>>

EVIOCGBIT

pub fn misc_mask(&self) -> Result<Bitmask<MiscKind>>

EVIOCGMASK

pub fn set_misc_mask(&self, bitmask: &Bitmask<MiscKind>) -> Result<()>

EVIOCSMASK

pub fn misc_bits(&self) -> Result<Bitmask<MiscKind>>

EVIOCGBIT

pub fn switch_mask(&self) -> Result<Bitmask<SwitchKind>>

EVIOCGMASK

pub fn set_switch_mask(&self, bitmask: &Bitmask<SwitchKind>) -> Result<()>

EVIOCSMASK

pub fn switch_bits(&self) -> Result<Bitmask<SwitchKind>>

EVIOCGBIT

pub fn led_mask(&self) -> Result<Bitmask<LedKind>>

EVIOCGMASK

pub fn set_led_mask(&self, bitmask: &Bitmask<LedKind>) -> Result<()>

EVIOCSMASK

pub fn led_bits(&self) -> Result<Bitmask<LedKind>>

EVIOCGBIT

pub fn sound_mask(&self) -> Result<Bitmask<SoundKind>>

EVIOCGMASK

pub fn set_sound_mask(&self, bitmask: &Bitmask<SoundKind>) -> Result<()>

EVIOCSMASK

pub fn sound_bits(&self) -> Result<Bitmask<SoundKind>>

EVIOCGBIT

pub fn autorepeat_mask(&self) -> Result<Bitmask<AutorepeatKind>>

EVIOCGMASK

pub fn set_autorepeat_mask(
&self,
bitmask: &Bitmask<AutorepeatKind>
) -> Result<()>

EVIOCSMASK

pub fn autorepeat_bits(&self) -> Result<Bitmask<AutorepeatKind>>

EVIOCGBIT

pub fn event_mask_raw(&self, kind: EventKind, buffer: &mut [u8]) -> Result<()>

EVIOCGMASK

pub fn set_event_mask_raw(&self, kind: EventKind, buffer: &[u8]) -> Result<()>

EVIOCSMASK

pub fn event_bits_raw(
&self,
kind: EventKind,
buffer: &mut [u8]
) -> Result<usize>

EVIOCGBIT

pub fn absolute_info(&self, abs: AbsoluteAxis) -> Result<AbsoluteInfo>

EVIOCGABS

pub fn set_absolute_info(
&self,
abs: AbsoluteAxis,
info: &AbsoluteInfo
) -> Result<()>

EVIOCSABS

pub fn grab(&self, grab: bool) -> Result<()>

EVIOCGRAB

pub fn revoke(&self) -> Result<()>

EVIOCREVOKE

pub fn set_clock_id(&self, value: i32) -> Result<()>

EVIOCSCLOCKID