Struct input_linux::uinput::UInputHandle

pub struct UInputHandle(_);

A handle to a uinput allowing the use of ioctls

Ownership of the file descriptor is not transferred, and it must stay open for this object's lifetime. It will not be closed automatically.

Methods

impl UInputHandle

pub fn new<F: AsRawFd>(fd: &F) -> Self

Create a new handle using an existing open file object.

pub fn from_fd(fd: RawFd) -> Self

Create a new handle from a raw file descriptor.

pub fn create_legacy(
&self,
id: &InputId,
name: &[u8],
ff_effects_max: u32,
abs: &[AbsoluteInfoSetup]
) -> Result<()>

Create a new uinput device using the legacy UI_DEV_CREATE interface

pub fn create(
&self,
id: &InputId,
name: &[u8],
ff_effects_max: u32,
abs: &[AbsoluteInfoSetup]
) -> Result<()>

Create a new uinput device, and fall back on the legacy interface if necessary

pub fn write(&self, events: &[input_event]) -> Result<usize>

Write input events to the device

pub fn read(&self, events: &mut [input_event]) -> Result<usize>

Read events from uinput (see EV_UINPUT)

pub fn sys_path(&self) -> Result<PathBuf>

pub fn evdev_name(&self) -> Result<OsString>

pub fn evdev_path(&self) -> Result<PathBuf>

pub fn dev_create(&self) -> Result<()>

UI_DEV_CREATE

pub fn dev_destroy(&self) -> Result<()>

UI_DEV_DESTROY

pub fn dev_setup(&self, value: &uinput_setup) -> Result<()>

UI_DEV_SETUP

pub fn abs_setup(&self, value: &uinput_abs_setup) -> Result<()>

UI_ABS_SETUP

pub fn set_evbit(&self, value: EventKind) -> Result<()>

UI_SET_EVBIT

pub fn set_keybit(&self, value: Key) -> Result<()>

UI_SET_KEYBIT

pub fn set_relbit(&self, value: RelativeAxis) -> Result<()>

UI_SET_RELBIT

pub fn set_absbit(&self, value: AbsoluteAxis) -> Result<()>

UI_SET_ABSBIT

pub fn set_mscbit(&self, value: MiscKind) -> Result<()>

UI_SET_MSCBIT

pub fn set_ledbit(&self, value: LedKind) -> Result<()>

UI_SET_LEDBIT

pub fn set_sndbit(&self, value: SoundKind) -> Result<()>

UI_SET_SNDBIT

pub fn set_ffbit(&self, value: i32) -> Result<()>

UI_SET_FFBIT

pub fn set_phys(&self, value: &CStr) -> Result<()>

UI_SET_PHYS

pub fn set_swbit(&self, value: SwitchKind) -> Result<()>

UI_SET_SWBIT

pub fn set_propbit(&self, value: InputProperty) -> Result<()>

UI_SET_PROPBIT

pub fn ff_upload_begin(&self, value: &mut uinput_ff_upload) -> Result<()>

UI_BEGIN_FF_UPLOAD

pub fn ff_upload_end(&self, value: &uinput_ff_upload) -> Result<()>

UI_END_FF_UPLOAD

pub fn ff_erase_begin(&self, value: &mut uinput_ff_erase) -> Result<()>

UI_BEGIN_FF_ERASE

pub fn ff_erase_end(&self, value: &uinput_ff_erase) -> Result<()>

UI_END_FF_ERASE

pub fn sys_name_buf(&self, buffer: &mut [u8]) -> Result<usize>

UI_GET_SYSNAME

pub fn sys_name(&self) -> Result<Vec<u8>>

UI_GET_SYSNAME

pub fn version(&self) -> Result<u32>

UI_GET_VERSION